+++ R/read_flights.R
@@ -68,6 +68,9 @@ if (length(date) == 1 & nchar(date[1])==6) {
   # check if download failed
   if (is.null(dt)) { return(invisible(NULL)) }
 
+  # convert columns to numeric
+  convert_to_numeric(dt)
+
   return(dt)
 }
 
@@ -128,6 +131,10 @@ pbapply::pboptions(original_options)
 
 # row bind data tables
 dt <- data.table::rbindlist(dt_list)
+
+# convert columns to numeric
+convert_to_numeric(dt)
+
 return(dt)
 }