+++ tests/testthat/test_read_flights.R
@@ -14,6 +14,7 @@ test_that("read_flights", {
   test1 <- read_flights()
   testthat::expect_true(is(test1, "data.table"))
   testthat::expect_true(nrow(test1) >0 )
+  testthat::expect_equal( class(test1$nr_voo), 'numeric')
 
   # one month, combinada, no progress bar
   test2 <- read_flights(date=202201, type='combinada' ,showProgress = FALSE)
@@ -33,7 +34,7 @@ test_that("read_flights", {
   testthat::expect_true(is(test4, "data.table"))
   testthat::expect_true(nrow(test4) >0 )
   months <- unique(test4$nr_mes_referencia)
-  testthat::expect_true( all.equal(months , c('1','2')) )
+  testthat::expect_true( all.equal(months , c(1,2)) )
 
  })